# improve List's indexer Set performance by 10% by doing array bound check on our own