@jkotas 2020-12-16
Do you have rational explanation for why this is faster?

There are two more instructions in the "after" code. Everything else being equal the "after" code should be slower.

I think you are likely just measuring some micro-architecture quirk that is not durable.