@EgorBo 2020-12-16
> Do you have rational explanation for why this is faster?
> 
> There are two more instructions in the hot path in the "after" code. Everything else being equal the "after" code should be slower.
> 
> I think you are likely just measuring some micro-architecture quirk that is not durable.

+, last time I tried to do the same trick for both setter/getter as part of https://github.com/dotnet/runtime/issues/45206 I decided to leave it as is because my benchmarks weren't happy about this change (but maybe I only tested getter, I don't remember)