@benaadams 2020-12-16
> There are two more instructions in the hot path in the "after" code. Everything else being equal the "after" code should be slower.

Looks to remove `sub` and `add` at entry and exit; and do a register cmp `r9d` rather than `rax+8`? Otherwise mostly the same (other than being a little shorter in the throw path, as only has one site)

```diff
-   sub rsp, 0x28

+   mov rax, [rcx+8]
    cmp edx, [rcx+0x10]
-   jae short L0022
+   jae short L001e
-   mov rax, [rcx+8]
+   mov r9d, [rax+8]

-   cmp edx, [rax+8]
+   cmp r9d, edx

-   jae short L0028
+   jbe short L001e

    movsxd rdx, edx
    mov [rax+rdx*4+0x10], r8d
    inc dword ptr [rcx+0x14]

-   add rsp, 0x28

    ret

+   mov rax, ThrowHelper.ThrowArgumentOutOfRange_IndexException()
+   jmp rax
-   call ThrowHelper.ThrowArgumentOutOfRange_IndexException()
-   int3
-   call 0x00007ffae0a0fc60
-   int3
```