@benaadams 2020-12-16
Also earlier load of `mov rax, [rcx+8]` I suppose giving more time before use?