@EgorBo 2020-12-16
Self-contained benchmark:
https://gist.github.com/EgorBo/a141bb171d09779f337d831aa763b43d

|              Method | value |      Mean |     Error |    StdDev |
|-------------------- |------ |----------:|----------:|----------:|
|    SetValue_int_old |   100 |  9.219 ns | 0.0041 ns | 0.0039 ns |
| SetValue_string_old |   100 | 19.379 ns | 0.0106 ns | 0.0099 ns |
|    SetValue_int_new |   100 |  9.308 ns | 0.0047 ns | 0.0044 ns |
| SetValue_string_new |   100 | 19.531 ns | 0.0319 ns | 0.0299 ns |

the numbers are different if I set different iteration counts.