@adamsitnik 2020-12-16
Thank you all for your review and feedback. Based on it, I am closing this PR.