@eiriktsarpalis 2020-12-16
Should we close the original as well?