@adamsitnik 2020-12-16
> Should we close the original issue as well?

I would get an answer from @BruceForstall to this question:

> perhaps you can dig into what might unblock hoisting of the array field fetch in the list case (see sharplab link above for the repro)?

If Bruce says there is nothing to improve or it would be too expensive, then we should close it.