@benaadams 2020-12-17
If you are looping (as per the benchmark) you can hoist the bounds check by converting the `List<T>` to a `Span<T>`; converting first to second
```csharp
public List<T> List()
{
    var list = _list;
    for (int i = 0; i < list.Count; i++)
        list[i] = default;
    return list;
}
        
public List<T> ListAsSpan()
{
    var span = CollectionsMarshal.AsSpan(_list);
    for (int i = 0; i < span.Count; i++)
        span[i] = default;
    return _list;
}
```